requirements.txt
selenium
pytest-selenium
allure-pytest
pytest-reportlog 
pyautogui
pyscreeze
pillow
gnome-screenshot 
opencv-python
pyperclip
pytest-bdd